afpacket: fix NewPacketSource params GOOS!=linux
The signature of `afpacket.NewPacketSource` was updated in
`readwriter.go` (`29ca59d5`) for `GOOS=linux`, but the signature for
that same function in `readwriter_other.go` for `GOOS!=linux` was not
updated to match. This causes a build failure when `GOOS!=linux`. The
build failure is not critical since only `GOOS=linux` is supported .